// src/models/appointment.model.ts
export interface Appointment {
    portName: string;
    address: string;
    city: string;
    state: string;
    country: string;
    plateNo: string;
    phoneNumber: string;
    trCompanyName: string;
    gstNo: string;
    transportLicNo: string;
    driverName: string;
    moveType: string;
    containerNumber: string;
    sizeType: string;
    line: string;
    chassisNo: string;
    appointmentStatus: string;
    appointmentCreated: string;
    appointmentValidThrough: string;
    appointmentLastModified: string;
    gateCode: string;
  }
  